+++ pkg/restore/createManager.go
@@ -25,7 +25,6 @@ type defaultCreateManager struct {
 	cleanup               cleanupManager
 	recorder              eventRecorder
 	maintenanceModeSwitch maintenanceModeSwitch
-	ownerRefRestorer      ownerReferenceRestore
 }
 
 func newCreateManager(
@@ -34,7 +33,6 @@ func newCreateManager(
 	namespace string,
 	recorder eventRecorder,
 	cleanup cleanupManager,
-	ownerRefRestorer ownerReferenceRestore,
 ) *defaultCreateManager {
 	maintenanceSwitch := repository.NewMaintenanceModeAdapter("k8s-backup-operator", ecosystemClientSet.CoreV1().ConfigMaps(namespace))
 	return &defaultCreateManager{
@@ -44,7 +42,6 @@ func newCreateManager(
 		recorder:              recorder,
 		maintenanceModeSwitch: maintenanceSwitch,
 		cleanup:               cleanup,
-		ownerRefRestorer:      ownerRefRestorer,
 	}
 }
 
@@ -108,11 +105,6 @@ func (cm *defaultCreateManager) create(ctx context.Context, restore *v1.Restore)
 	err = provider.SyncBackups(ctx)
 	if err != nil {
 		return fmt.Errorf("failed to sync backups with provider: %w", err)
-	}
-
-	err = cm.ownerRefRestorer.RestoreOwnerReferences(ctx)
-	if err != nil {
-		return fmt.Errorf("failed to restore owner references: %w", err)
 	}
 
 	_, err = restoreClient.UpdateStatusCompleted(ctx, restore)