+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v2.2.3] - 2026-01-28
+### Removed
+- [#97], [#89] Backup & restore of owner references
+  - This is not needed anymore, because we only back up Dogu-specific data.
+    The dogu-operator is able to restore owner references itself now.
+
 ## [v2.2.2] - 2026-01-22
 ### Changed
 - [#93] Update backup documentation